requirements.txt
asgiref==3.9.2
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
joblib==1.5.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.11.0
xgboost==3.1.1
